#73 fix broken tests
